+++ pkg/querier/tripperware/queryrange/query_range.go
@@ -41,12 +41,43 @@ var (
 	cacheControlHeader = "Cache-Control"
 )
 
-type prometheusCodec struct {
-	sharded bool
+func (resp *PrometheusResponse) HTTPHeaders() map[string][]string {
+	if resp != nil && resp.GetHeaders() != nil {
+		r := map[string][]string{}
+		for _, header := range resp.GetHeaders() {
+			if header != nil {
+				r[header.Name] = header.Values
+			}
+		}
+
+		return r
+	}
+	return nil
 }
 
-func NewPrometheusCodec(sharded bool) *prometheusCodec { //nolint:revive
-	return &prometheusCodec{sharded: sharded}
+type prometheusCodec struct {
+	tripperware.Codec
+	sharded          bool
+	compression      tripperware.Compression
+	defaultCodecType tripperware.CodecType
+}
+
+func NewPrometheusCodec(sharded bool, compressionStr string, defaultCodecTypeStr string) *prometheusCodec { //nolint:revive
+	compression := tripperware.NonCompression // default
+	if compressionStr == string(tripperware.GzipCompression) {
+		compression = tripperware.GzipCompression
+	}
+
+	defaultCodecType := tripperware.JsonCodecType // default
+	if defaultCodecTypeStr == string(tripperware.ProtobufCodecType) {
+		defaultCodecType = tripperware.ProtobufCodecType
+	}
+
+	return &prometheusCodec{
+		sharded:          sharded,
+		compression:      compression,
+		defaultCodecType: defaultCodecType,
+	}
 }
 
 func (c prometheusCodec) MergeResponse(ctx context.Context, req tripperware.Request, responses ...tripperware.Response) (tripperware.Response, error) {
@@ -57,6 +88,11 @@ func (c prometheusCodec) MergeResponse(ctx context.Context, req tripperware.Requ
 		return tripperware.NewEmptyPrometheusResponse(false), nil
 	}
 
+	// Safety guard in case any response from results cache middleware
+	// still uses the old queryrange.PrometheusResponse type.
+	for i, resp := range responses {
+		responses[i] = convertToTripperwarePrometheusResponse(resp)
+	}
 	return tripperware.MergeResponse(ctx, c.sharded, nil, responses...)
 }
 
@@ -116,7 +152,7 @@ func (c prometheusCodec) DecodeRequest(_ context.Context, r *http.Request, forwa
 	return &result, nil
 }
 
-func (prometheusCodec) EncodeRequest(ctx context.Context, r tripperware.Request) (*http.Request, error) {
+func (c prometheusCodec) EncodeRequest(ctx context.Context, r tripperware.Request) (*http.Request, error) {
 	promReq, ok := r.(*tripperware.PrometheusRequest)
 	if !ok {
 		return nil, httpgrpc.Errorf(http.StatusBadRequest, "invalid request format")
@@ -140,8 +176,7 @@ func (prometheusCodec) EncodeRequest(ctx context.Context, r tripperware.Request)
 		}
 	}
 
-	// Always ask gzip to the querier
-	h.Set("Accept-Encoding", "gzip")
+	tripperware.SetRequestHeaders(h, c.defaultCodecType, c.compression)
 
 	req := &http.Request{
 		Method:     "GET",
@@ -154,7 +189,7 @@ func (prometheusCodec) EncodeRequest(ctx context.Context, r tripperware.Request)
 	return req.WithContext(ctx), nil
 }
 
-func (prometheusCodec) DecodeResponse(ctx context.Context, r *http.Response, _ tripperware.Request) (tripperware.Response, error) {
+func (c prometheusCodec) DecodeResponse(ctx context.Context, r *http.Response, _ tripperware.Request) (tripperware.Response, error) {
 	log, ctx := spanlogger.New(ctx, "ParseQueryRangeResponse") //nolint:ineffassign,staticcheck
 	defer log.Finish()
 
@@ -173,8 +208,19 @@ func (prometheusCodec) DecodeResponse(ctx context.Context, r *http.Response, _ t
 	log.LogFields(otlog.Int("bytes", len(buf)))
 
 	var resp tripperware.PrometheusResponse
-	if err := json.Unmarshal(buf, &resp); err != nil {
+	err = tripperware.UnmarshalResponse(r, buf, &resp)
+
+	if err != nil {
 		return nil, httpgrpc.Errorf(http.StatusInternalServerError, "error decoding response: %v", err)
+	}
+
+	// protobuf serialization treats empty slices as nil
+	if resp.Data.ResultType == model.ValMatrix.String() && resp.Data.Result.GetMatrix().SampleStreams == nil {
+		resp.Data.Result.GetMatrix().SampleStreams = []tripperware.SampleStream{}
+	}
+
+	if resp.Headers == nil {
+		resp.Headers = []*tripperware.PrometheusResponseHeader{}
 	}
 
 	for h, hv := range r.Header {
@@ -206,7 +252,7 @@ func (prometheusCodec) EncodeResponse(ctx context.Context, res tripperware.Respo
 
 	resp := http.Response{
 		Header: http.Header{
-			"Content-Type": []string{"application/json"},
+			"Content-Type": []string{tripperware.ApplicationJson},
 		},
 		Body:          io.NopCloser(bytes.NewBuffer(b)),
 		StatusCode:    http.StatusOK,